refactor: extract mqtt_demo/ — move bridge, config, logger, appliance descriptors
