+++ mqtt_demo/Dockerfile
@@ -3,12 +3,15 @@ FROM python:3.11-slim
 WORKDIR /app
 
 # Python deps first so layer cache survives code changes
-COPY requirements.txt .
+COPY mqtt_demo/requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 
-# Application code
-COPY main.py .
-COPY samsung_appliance/ ./samsung_appliance/
+# Application code — build context is the repo root (see
+# docker-compose.yml's `context: ..`) since mqtt_demo/ depends on the
+# protocol/ and ocf/ library packages that live alongside it.
+COPY protocol/ ./protocol/
+COPY ocf/ ./ocf/
+COPY mqtt_demo/ ./mqtt_demo/
 
 # /config holds the ab0b0ac4 client cert + key. Mount from the host so
 # secrets aren't baked into the image.
@@ -28,4 +31,4 @@ ENV CERT_PATH=/config/ab0b0ac4_fullchain.pem \
 
 # No port — bridge is outbound-only (DTLS UDP to appliance, MQTT to broker).
 
-CMD ["python", "main.py"]
+CMD ["python", "-m", "mqtt_demo"]
