--- samsung_appliance/__init__.py
@@ -1,5 +1,0 @@
-"""Samsung appliance local-API → MQTT bridge with HA discovery.
-
-Multi-device: dryer, oven, etc. The device class is selected at startup
-via the DEVICE_CLASS env var (see samsung_appliance.appliances)."""
-__version__ = "2.0.0"
